MIPS64: Fix sim Push/Pop Address.
temo sp variable was 32-bit int.

+++ src/mips64/simulator-mips64.cc
@@ -3022,7 +3022,7 @@ double Simulator::CallFP(byte* entry, double d0, double d1) {
 
 
 uintptr_t Simulator::PushAddress(uintptr_t address) {
-  int new_sp = get_register(sp) - sizeof(uintptr_t);
+  int64_t new_sp = get_register(sp) - sizeof(uintptr_t);
   uintptr_t* stack_slot = reinterpret_cast<uintptr_t*>(new_sp);
   *stack_slot = address;
   set_register(sp, new_sp);
@@ -3031,7 +3031,7 @@ uintptr_t Simulator::PushAddress(uintptr_t address) {
 
 
 uintptr_t Simulator::PopAddress() {
-  int current_sp = get_register(sp);
+  int64_t current_sp = get_register(sp);
   uintptr_t* stack_slot = reinterpret_cast<uintptr_t*>(current_sp);
   uintptr_t address = *stack_slot;
   set_register(sp, current_sp + sizeof(uintptr_t));
